move getHrefLink inline
